requirements.txt
accelerate==1.13.0
bitsandbytes==0.49.2
datasets==4.8.3
diffusers==0.37.0
huggingface_hub==1.7.2
numpy==2.4.3
packaging==26.0
peft==0.18.1
Pillow==12.1.1
prodigyopt==1.1.2
safetensors==0.7.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==5.3.0
wandb==0.25.1
xformers==0.0.35
